Reorganize docker.transport package
