+++ docker/transport/__init__.py
@@ -1,3 +1,6 @@
 # flake8: noqa
-from .npipe import NpipeAdapter
 from .unixconn import UnixAdapter
+try:
+    from .npipeconn import NpipeAdapter
+except ImportError:
+    pass
