--- docker/transport/npipe/__init__.py
@@ -1,1 +1,0 @@
-from .npipeconn import NpipeAdapter  # flake8: noqa
