+++ docker/transport/npipesocket.py
