--- docker/transport/unixconn/__init__.py
@@ -1,1 +1,0 @@
-from .unixconn import UnixAdapter  # flake8: noqa
